Move extension functions for `Query` to new file
It was in `DatabaseReferen.kt` and is now in `Query.kt`.
