+++ integration/kotlinx-coroutines-firebase-android/kotlinx-coroutines-firebase-android/src/main/java/kotlinx/coroutines/experimental/firebase/android/DatabaseReference.kt
@@ -208,47 +208,3 @@ public suspend fun <T : Any> DatabaseReference.saveValue(value: T): Unit =
         setValue(value).await().let { kotlin.Unit }
 
 public suspend fun <T : Any> DatabaseReference.pushValue(value: T): Unit = push().saveValue(value)
-
-private suspend fun <T : Any> awaitQuerySingleValue(query: Query, type: Class<T>): T =
-        suspendCancellableCoroutine { continuation ->
-            val listener = object : ValueEventListener {
-                override fun onDataChange(snapshot: DataSnapshot) = try {
-                    continuation.resume(snapshot.getValue(type)!!)
-                } catch (exception: Exception) {
-                    continuation.resumeWithException(exception)
-                }
-
-                override fun onCancelled(error: DatabaseError) =
-                        continuation.resumeWithException(error.toException())
-            }
-
-            query.addListenerForSingleValueEvent(listener)
-            continuation.invokeOnCompletion { query.removeEventListener(listener) }
-        }
-
-public suspend fun <T : Any> Query.readValue(type: Class<T>): T = awaitQuerySingleValue(this, type)
-
-public suspend inline fun <reified T : Any> Query.readValue(): T = readValue(T::class.java)
-
-private suspend fun <T : Any> awaitQueryListValue(query: Query, type: Class<T>): List<T> =
-        suspendCancellableCoroutine { continuation ->
-            val listener = object : ValueEventListener {
-                override fun onDataChange(snapshot: DataSnapshot) = try {
-                    val data: List<T> = snapshot.children.toHashSet().map { it.getValue(type)!! }
-
-                    continuation.resume(data)
-                } catch (exception: Exception) {
-                    continuation.resumeWithException(exception)
-                }
-
-                override fun onCancelled(error: DatabaseError) =
-                        continuation.resumeWithException(error.toException())
-            }
-
-            query.addListenerForSingleValueEvent(listener)
-            continuation.invokeOnCompletion { query.removeEventListener(listener) }
-        }
-
-public suspend fun <T : Any> Query.readList(type: Class<T>): List<T> = awaitQueryListValue(this, type)
-
-public suspend inline fun <reified T : Any> Query.readList(): List<T> = readList(T::class.java)
